Update AgentFactory implementations to handle existing agents

+++ dotnet/samples/GettingStartedWithAgents/OpenAIAssistant/Step07_Assistant_Declarative.cs
@@ -92,6 +92,26 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
     }
 
     /// <summary>
+    /// Demonstrates loading an existing OpenAI Assistant.
+    /// </summary>
+    [Fact]
+    public async Task OpenAIAssistantAgentWithId()
+    {
+        var text =
+            """
+            id: ${AzureOpenAI:AgentId}
+            type: openai_assistant
+            name: StoryAgent
+            instructions: Tell a story suitable for children about the topic provided by the user. You always respond in French.
+            """;
+        OpenAIAssistantAgentFactory factory = new();
+
+        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Kernel = this._kernel }, configuration: TestConfiguration.ConfigurationRoot);
+
+        await InvokeAgentAsync(agent!, "Cats and Dogs", false);
+    }
+
+    /// <summary>
     /// Demonstrates creating and using a OpenAI Assistant with templated instructions.
     /// </summary>
     [Fact]
@@ -159,6 +179,11 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
 
     public Step07_Assistant_Declarative(ITestOutputHelper output) : base(output)
     {
+        // Uncomment the following lines to enable logging
+        //var handler = new LoggingHandler(new HttpClientHandler(), output);
+        //var httpClient = new HttpClient(handler);
+        //var client = OpenAIAssistantAgent.CreateAzureOpenAIClient(new AzureCliCredential(), new Uri(this.Endpoint!), httpClient);
+
         var builder = Kernel.CreateBuilder();
         builder.Services.AddSingleton<OpenAIClient>(this.Client);
         this._kernel = builder.Build();
@@ -170,7 +195,7 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
     /// <summary>
     /// Invoke the agent with the user input.
     /// </summary>
-    private async Task InvokeAgentAsync(Agent agent, string input)
+    private async Task InvokeAgentAsync(Agent agent, string input, bool deleteAgent = true)
     {
         AgentThread? agentThread = null;
         try
@@ -187,9 +212,12 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
         }
         finally
         {
-            var openaiAgent = agent as OpenAIAssistantAgent;
-            Assert.NotNull(openaiAgent);
-            await openaiAgent.Client.DeleteAssistantAsync(openaiAgent.Id);
+            if (deleteAgent)
+            {
+                var openaiAgent = agent as OpenAIAssistantAgent;
+                Assert.NotNull(openaiAgent);
+                await openaiAgent.Client.DeleteAssistantAsync(openaiAgent.Id);
+            }
 
             if (agentThread is not null)
             {
